Update README.md
Hello everyone, first I would like to say that I think it is a great shame that EmperorsToolsShop has stopped. Many of you are probably now without tools to get SMTPs and leads because the licenses have been removed. I bet nobody has the source because he asked a absurd amount for that.

I offer a package of tools that are NOT licensed so once you have the file you just have a lifetime, no bullshitting with license keys. Some files are in .py, some files are .exe

Why am I selling these tools? Because it's a shitty thing that it just stopped out of nowhere, leaving a lot of people without tools to grab SMTPs and because of course I can still earn something from it. I have checked all the tools and they all still work perfectly. After this post I will give a separate explanation for each tool about how it works and what it does. 

For any questions you can PM me on telegram on: @Multireseller. Only serious people, people that talk shit i will block directly since im only here to help and not to discuss.

DISCLAIMER: IM NOT THE OFFICIAL EMPERORSTOOLSSHOP IM JUST A RESELLER.